Separated out configs and added sass loaders
